+++ headers/interactives.h
@@ -17,14 +17,29 @@ struct Sign : public Object<RectCollider,TextureRenderer,Sign>
     size_t getMessagesSize();
 };
 
+struct Rover : public Object<RectCollider,TextureRenderer,Rover>
+{
+    static constexpr Vector2 ROVER_DIMEN = {110,110};
+    static constexpr std::string ROVER_SPRITE_PATH = "rover.png";
+    Rover();
+    void interactWith(PhysicsBody& other);
+};
+
 template<>
 struct Factory<Sign>
 {
     static constexpr std::string ObjectName = "sign";
     using Base = FactoryBase<Sign,
                                 access<Sign,&Sign::message>,
-                                access<Sign,&Sign::orient,&Orient::pos>,
-                                access<Sign,&Sign::orient,&Orient::layer>>;
+                                access<Sign,&Sign::orient,&Orient::pos>>;
+};
+
+template<>
+struct Factory<Rover>
+{
+    static constexpr std::string ObjectName = "rover";
+    using Base = FactoryBase<Rover,
+                                access<Rover,&Rover::orient,&Orient::pos>>;
 };
 
 #endif // INTERACTIVES_H_INCLUDED