+++ headers/portal.h
@@ -23,6 +23,7 @@ struct TokenLocked : public PortalCondition
     size_t requirement = 0;
     TokenLocked(size_t req) : requirement(req)
     {
+
     }
     bool unlocked();
     void render(Shape shape);
@@ -35,10 +36,10 @@ struct TokenLocked : public PortalCondition
 struct Portal : public Object<CircleCollider,ShapeRenderer<CIRCLE>,Portal>
 {
     std::unique_ptr<PortalCondition> cond;
-    Orient dest;
+    Vector2 destPos = {};
+    int layerDisp = 0;
     static Shader PortalShader;
     RenderTexture2D texture;
-    Portal(const Vector3& start, int radius, const Vector3& dest_);
     Portal();
     bool unlocked();
     void update()
@@ -49,82 +50,16 @@ struct Portal : public Object<CircleCollider,ShapeRenderer<CIRCLE>,Portal>
     void render();
 };
 
-//spawn a portal when interacted with
-struct TriggerPortalSpawn : public InteractComponent
-{
-    bool active = true;
-    bool transition = false;
-    bool absolute = true; //true if "start" is the absolute position to spawn the portal; false if instead start is the displacement from current position
-
-    Vector3 start = {}; //portal start point
-    Vector3 end = {}; //portal end point
-
-    Key::KeyVal lockVal = Key::UNLOCKED;
-
-    static InteractAction createInteractFunc();
-
-    TriggerPortalSpawn(bool transition_, const Vector3& start_,
-                       const Vector3& end_, int radius, Key::KeyVal keyVal_ = Key::UNLOCKED) :
-                                                                transition(transition_),
-                                                                lockVal(keyVal_),
-                                                                InteractComponent({}),
-                                                                start(start_),
-                                                                end(end_)
-    {
-
-    }
-
-    TriggerPortalSpawn(bool transition_, const Vector2& disp,
-                       const Vector3& end_, int radius, Key::KeyVal keyVal_ = Key::UNLOCKED) :
-                                                        transition(transition_),
-                                                        lockVal(keyVal_),
-                                                        InteractComponent({}),
-                                                        start(disp.x,disp.y,0),
-                                                        end(end_),
-                                                        absolute(false)
-
-    {
-
-    }
-
-    TriggerPortalSpawn() : InteractComponent({})
-    {
-
-    }
-
-    void interact(PhysicsBody&, PhysicsBody& other);
-
-};
-
-//using PortalSpawner = Object<CircleCollider,TextureRenderer,TriggerPortalSpawn>;
-
 template<>
 struct Factory<Portal>
 {
     static constexpr std::string ObjectName = "portal";
     using Base = FactoryBase<Portal,
-                                access<Portal,&Portal::dest,&Orient::pos>,
-                                access<Portal,&Portal::dest,&Orient::layer>,
+                                access<Portal,&Portal::destPos>,
+                                access<Portal,&Portal::layerDisp>,
                                 access<Portal,&Portal::orient,&Orient::pos>,
-                                access<Portal,&Portal::orient,&Orient::layer>,
                                 access<Portal,&Portal::cond>,
                                 access<Portal,&Portal::collider,&CircleCollider::radius>>;
 };
-
-/*template<>
-struct Factory<PortalSpawner>
-{
-    static constexpr std::string ObjectName = "portal_spawner";
-    using Base = FactoryBase<PortalSpawner,
-                                access<PortalSpawner,&PortalSpawner::orient,&Orient::pos>,
-                                access<PortalSpawner,&PortalSpawner::orient,&Orient::layer>,
-                                access<PortalSpawner,&PortalSpawner::collider,&CircleCollider::radius>,
-                                access<PortalSpawner,&PortalSpawner::renderer,&TextureRenderer::sprite>,
-                                access<PortalSpawner,&PortalSpawner::collideTrigger,&TriggerPortalSpawn::start>,
-                                access<PortalSpawner,&PortalSpawner::collideTrigger,&TriggerPortalSpawn::end>,
-                                access<PortalSpawner,&PortalSpawner::collideTrigger,&TriggerPortalSpawn::absolute>>;
-
-};*/
-
 
 #endif // PORTAL_H_INCLUDED