+++ src/interface.cpp
@@ -62,7 +62,7 @@ void Interface::render()
                        0,it);
         i ++;
     }
-    float height = Globals::Game.player->power/Player::PLAYER_MAX_POWER*100;
+    float height = Globals::Game.player->dying;
     DrawRectangle(100,screenDimen.x*.9 - height,100,height,RED);
 
     DrawText(std::to_string(Globals::Game.getCollects()).c_str(),.85*screenDimen.x,.9*screenDimen.y,50,WHITE);